Replaced tests.
