+++ bootstrap.js
@@ -9,8 +9,6 @@ import { ChildProcess } from 'node:child_process';
 import _ from 'underscore';
 import axios from 'axios';
 import op from 'object-path';
-import config from './config.json' assert { type: 'json' };
-import packagejson from './package.json' assert { type: 'json' };
 import chalk from 'chalk';
 import fs from 'fs-extra';
 import { globbySync as globby } from './globby-patch.js';
@@ -32,6 +30,12 @@ import prompt from 'prompt';
 import decompress from '@atomic-reactor/decompress';
 import { error, message } from './lib/messenger.js';
 
+const normalizePath = (...args) => path.normalize(path.join(...args));
+
+const root = path.dirname(fileURLToPath(import.meta.url));
+const config = fs.readJsonSync(normalizePath(root, 'config.json'));
+const packagejson = fs.readJsonSync(normalizePath(root, 'package.json'));
+
 const importer = (path, options) => {
     options = options || {}; 
 
@@ -48,7 +52,7 @@ const bootstrap = async () => {
     const cwd = path.resolve(process.cwd());
     const homedir = os.homedir();
     const ver = op.get(packagejson, 'version');
-    const root = path.dirname(fileURLToPath(import.meta.url));
+    
 
     global.Actinium = Actinium;
     global.Hook = Hook;
@@ -92,7 +96,7 @@ const bootstrap = async () => {
             return output;
         }, {});
 
-    const normalizePath = (...args) => path.normalize(path.join(...args));
+    
 
     const initialize = async props => {
         // import arlic-hooks.js files
@@ -186,6 +190,7 @@ const bootstrap = async () => {
         decompress,
         deleteEmpty,
         error,
+        fileURLToPath,
         flagsToParams,
         fs,
         path,
@@ -219,6 +224,12 @@ const bootstrap = async () => {
 
     // Build the props object
     global.arcli.props = await initialize(global.arcli.props);
+
+    global.arcli.props.inquirer = inquirer;
+    global.arcli.props.prompt = prompt;
+    global.arcli.prefix = chalk.cyan(
+        String(arcli.props.config.prompt.prefix).trim(),
+    );
 
     return global.arcli;
 };
