+++ package.json
@@ -1,91 +1,85 @@
 {
-  "name": "@atomic-reactor/cli",
-  "version": "3.0.0",
-  "description": "A CLI for creating Atomic Reactor Awesomeness",
-  "type": "module",
-  "scripts": {
-    "test": "run-s test:*",
-    "test:pg": "mocha ./test/blah.spec.js",
-    "test:reactium-install": "mocha ./test/reactium-install.spec.js",
-    "test:config-set": "mocha ./test/config-set.spec.js",
-    "test:reactium-update": "mocha ./test/reactium-update.spec.js",
-    "test:project-init": "mocha ./test/project-init.spec.js",
-    "test:glob": "mocha ./test/glob.spec.js",
-    "pub": "node ./prepublish.js && npm publish"
-  },
-  "author": "Reactium LLC",
-  "license": "MIT",
-  "engines": {
-    "node": ">=18.12.1",
-    "npm": ">=9.1.1"
-  },
-  "repository": {
-    "type": "git",
-    "url": "git+https://github.com/Atomic-Reactor/CLI.git"
-  },
-  "bin": {
-    "arcli": "arcli.js"
-  },
-  "keywords": [
-    "actinium",
-    "react",
-    "reactium",
-    "redux",
-    "framework",
-    "javascript",
-    "ddd",
-    "cli"
-  ],
-  "dependencies": {
-    "@atomic-reactor/decompress": "^4.2.5",
-    "@atomic-reactor/reactium-sdk-core": "^1.2.16",
-    "action-sequence": "^1.1.2",
-    "axios": "^1.1.3",
-    "camelcase": "^6.2.0",
-    "chalk": "^4.1.2",
-    "cli-spinners": "^2.7.0",
-    "commander": "^6.1.0",
-    "crypto": "^1.0.1",
-    "decamelize": "^4.0.0",
-    "delete-empty": "^3.0.0",
-    "folder-zipper": "^1.0.0",
-    "fs-extra": "^10.1.0",
-    "fs-readdir-recursive": "^1.1.0",
-    "globby": "^13.1.2",
-    "handlebars": "^4.7.7",
-    "ignored": "^2.0.4",
-    "inquirer": "^7.3.3",
-    "inquirer-autocomplete-prompt": "^1.0.2",
-    "inquirer-fuzzy-path": "^2.3.0",
-    "memory-cache": "^0.2.0",
-    "micromatch": "^4.0.5",
-    "moment": "^2.29.4",
-    "object-path": "^0.11.8",
-    "ora": "^5.1.0",
-    "parse": "^3.4.4",
-    "pm2": "^5.1.2",
-    "portscanner": "^2.2.0",
-    "prettier": "1.18.1",
-    "prompt": "^1.0.0",
-    "request": "^2.88.2",
-    "semver": "^7.3.8",
-    "slugify": "^1.4.5",
-    "tar": "^6.1.11",
-    "text-table": "^0.2.0",
-    "underscore": "^1.13.6",
-    "uuid": "^3.3.3"
-  },
-  "devDependencies": {
-    "chai": "^4.3.6",
-    "mocha": "^9.1.3",
-    "npm-run-all": "^4.1.5"
-  },
-  "bugs": {
-    "url": "https://github.com/Atomic-Reactor/CLI/issues"
-  },
-  "homepage": "https://github.com/Atomic-Reactor/CLI#readme",
-  "directories": {
-    "lib": "lib",
-    "test": "test"
-  }
+    "name": "@atomic-reactor/cli",
+    "version": "3.0.0",
+    "description": "A CLI for creating Atomic Reactor Awesomeness",
+    "type": "module",
+    "scripts": {
+        "test": "mocha ./test/init.spec.js",
+        "pub": "node ./prepublish.js && npm publish"
+    },
+    "author": "Reactium LLC",
+    "license": "MIT",
+    "engines": {
+        "node": ">=18.12.1",
+        "npm": ">=9.1.1"
+    },
+    "repository": {
+        "type": "git",
+        "url": "git+https://github.com/Atomic-Reactor/CLI.git"
+    },
+    "bin": {
+        "arcli": "arcli.js"
+    },
+    "keywords": [
+        "actinium",
+        "react",
+        "reactium",
+        "redux",
+        "framework",
+        "javascript",
+        "ddd",
+        "cli"
+    ],
+    "dependencies": {
+        "@atomic-reactor/decompress": "^4.2.5",
+        "@atomic-reactor/reactium-sdk-core": "^1.2.16",
+        "action-sequence": "^1.1.2",
+        "axios": "^1.1.3",
+        "camelcase": "^6.2.0",
+        "chalk": "^4.1.2",
+        "cli-spinners": "^2.7.0",
+        "commander": "^6.1.0",
+        "crypto": "^1.0.1",
+        "decamelize": "^4.0.0",
+        "delete-empty": "^3.0.0",
+        "folder-zipper": "^1.0.0",
+        "fs-extra": "^10.1.0",
+        "fs-readdir-recursive": "^1.1.0",
+        "globby": "^13.1.2",
+        "handlebars": "^4.7.7",
+        "ignored": "^2.0.4",
+        "inquirer": "^7.3.3",
+        "inquirer-autocomplete-prompt": "^1.0.2",
+        "inquirer-fuzzy-path": "^2.3.0",
+        "memory-cache": "^0.2.0",
+        "micromatch": "^4.0.5",
+        "moment": "^2.29.4",
+        "object-path": "^0.11.8",
+        "ora": "^5.1.0",
+        "parse": "^3.4.4",
+        "pm2": "^5.1.2",
+        "portscanner": "^2.2.0",
+        "prettier": "1.18.1",
+        "prompt": "^1.0.0",
+        "request": "^2.88.2",
+        "semver": "^7.3.8",
+        "slugify": "^1.4.5",
+        "tar": "^6.1.11",
+        "text-table": "^0.2.0",
+        "underscore": "^1.13.6",
+        "uuid": "^3.3.3"
+    },
+    "devDependencies": {
+        "chai": "^4.3.6",
+        "mocha": "^9.1.3",
+        "npm-run-all": "^4.1.5"
+    },
+    "bugs": {
+        "url": "https://github.com/Atomic-Reactor/CLI/issues"
+    },
+    "homepage": "https://github.com/Atomic-Reactor/CLI#readme",
+    "directories": {
+        "lib": "lib",
+        "test": "test"
+    }
 }
